chore(remix): remove remix and use vite-react-template
* with react-router v7, remix is all but dead and this example doesn't need routing
* remix was originally added to handle the meta tags
* I have since learnt that I gain nothing from that implementation
