+++ index.html
@@ -1,0 +1,22 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Redux TodoMVC Example</title>
+    <meta property="og:title" content="Redux TodoMVC Example" />
+    <meta
+      name="description"
+      content="TodoMVC example application written using React, Redux and Vite"
+    />
+    <meta
+      name="og:image"
+      content="https://upload.wikimedia.org/wikipedia/commons/5/5d/GNOME_Todo_icon_2019.svg"
+    />
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="module" src="/src/main.tsx"></script>
+  </body>
+</html>
