+++ package.json
@@ -12,7 +12,6 @@
     "@eslint/compat": "^1.2.5",
     "@ianvs/prettier-plugin-sort-imports": "^4.4.0",
     "@playwright/test": "^1.49.1",
-    "@remix-run/dev": "^2.15.2",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.1.0",
     "@testing-library/user-event": "^14.5.2",
@@ -57,13 +56,7 @@
     "vitest": "^2.1.8"
   },
   "dependencies": {
-    "@netlify/edge-functions": "^2.11.1",
-    "@netlify/remix-edge-adapter": "^3.4.2",
-    "@netlify/remix-runtime": "^2.3.1",
     "@reduxjs/toolkit": "^2.5.0",
-    "@remix-run/node": "^2.15.2",
-    "@remix-run/react": "^2.15.2",
-    "@remix-run/serve": "^2.15.2",
     "clsx": "^2.1.1",
     "isbot": "^5.1.21",
     "motion": "^11.17.0",
@@ -75,8 +68,8 @@
   },
   "scripts": {
     "prepare": "node .husky/install.mjs",
-    "dev": "remix vite:dev",
-    "build": "remix vite:build",
+    "dev": "vite",
+    "build": "tsc -b && vite build",
     "start": "netlify serve",
     "fmt": "prettier --write .",
     "fmt.check": "prettier --check .",
