--- src/entry.server.tsx
@@ -1,2 +1,0 @@
-// @ts-expect-error virtual module
-export { default } from "virtual:netlify-server-entry";
