+++ src/vite-env.d.ts
@@ -1,2 +1,1 @@
-/// <reference types="@remix-run/node" />
 /// <reference types="vite/client" />
